Convert bootstrap script to ESM
