+++ .changeset/ready-guests-hide.md
@@ -1,0 +1,5 @@
+---
+"roo-cline": patch
+---
+
+Convert bootstrap script to esm
